requirements.txt
pyDes
paho-mqtt==1.3.0
urllib
